+++ src/functionalTest/functional/uk/gov/hmcts/reform/em/hrs/functional/DownloadHearingRecordingScenarios.java
@@ -6,10 +6,10 @@ import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.em.hrs.functional.util.TestUtil;
+import uk.gov.hmcts.reform.em.hrs.functional.config.TestUtil;
+
 
 import java.util.Optional;
-import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import static org.hamcrest.CoreMatchers.is;
@@ -28,9 +28,8 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
         testUtil.deleteFileFromCvpContainer(FOLDER);
         testUtil.deleteFileFromHrsContainer(FOLDER);
 
-        final String id = UUID.randomUUID().toString();
-        FILE_NAME = String.format(FILE_NAME, id);
-        testUtil.uploadToCvpContainer(FILE_NAME);
+        fileName = String.format(fileName, counter.incrementAndGet());
+        testUtil.uploadToCvpContainer(fileName);
         createFolderIfDoesNotExistInHrsDB(FOLDER);
     }
 
@@ -42,7 +41,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
     @Test
     public void anUserWithCaseWorkerHrsRoleShouldBeAbleToDownloadHearingRecordings() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
+        final JsonNode segmentPayload = getSegmentPayload(fileName);
 
         postRecordingSegment(segmentPayload)
             .then()
@@ -51,7 +50,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
         TimeUnit.SECONDS.sleep(20);
 
-        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
+        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCcdByRecordingReference(fileName);
         assertTrue(optionalCaseDetails.isPresent());
 
         final CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
@@ -76,7 +75,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
     @Test
     public void anUserWithCaseWorkerRoleShouldNotBeAbleToDownloadHearingRecordings() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
+        final JsonNode segmentPayload = getSegmentPayload(fileName);
 
         postRecordingSegment(segmentPayload)
             .then()
@@ -85,7 +84,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
         TimeUnit.SECONDS.sleep(20);
 
-        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
+        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCcdByRecordingReference(fileName);
         assertTrue(optionalCaseDetails.isPresent());
 
         final CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
@@ -111,7 +110,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
     @Test
     public void anUserWithCitizenRoleShouldNotBeAbleToDownloadHearingRecordings() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
+        final JsonNode segmentPayload = getSegmentPayload(fileName);
 
         postRecordingSegment(segmentPayload)
             .then()
@@ -120,7 +119,7 @@ public class DownloadHearingRecordingScenarios extends BaseTest {
 
         TimeUnit.SECONDS.sleep(20);
 
-        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCCDByRecordingReference(FILE_NAME);
+        final Optional<CaseDetails> optionalCaseDetails = findCaseDetailsInCcdByRecordingReference(fileName);
         assertTrue(optionalCaseDetails.isPresent());
 
         final CaseDetails caseDetails = optionalCaseDetails.orElseGet(() -> CaseDetails.builder().build());
