+++ src/functionalTest/functional/uk/gov/hmcts/reform/em/hrs/functional/HearingRecordingSegmentScenarios.java
@@ -6,9 +6,9 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import uk.gov.hmcts.reform.em.hrs.functional.util.TestUtil;
+import uk.gov.hmcts.reform.em.hrs.functional.config.TestUtil;
 
-import java.util.UUID;
+
 import java.util.concurrent.TimeUnit;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -20,14 +20,13 @@ public class HearingRecordingSegmentScenarios extends BaseTest {
     @Autowired
     private TestUtil testUtil;
 
+
     @Before
     public void setup() throws Exception {
         testUtil.deleteFileFromHrsContainer(FOLDER);
         testUtil.deleteFileFromCvpContainer(FOLDER);
-
-        final String id = UUID.randomUUID().toString();
-        FILE_NAME = String.format(FILE_NAME, id);
-        testUtil.uploadToCvpContainer(FILE_NAME);
+        fileName = String.format(fileName, counter.incrementAndGet());
+        testUtil.uploadToCvpContainer(fileName);
         createFolderIfDoesNotExistInHrsDB(FOLDER);
     }
 
@@ -39,7 +38,7 @@ public class HearingRecordingSegmentScenarios extends BaseTest {
 
     @Test
     public void shouldCreateHearingRecordingSegment() throws Exception {
-        final JsonNode segmentPayload = getSegmentPayload(FILE_NAME);
+        final JsonNode segmentPayload = getSegmentPayload(fileName);
 
         postRecordingSegment(segmentPayload)
             .then()
@@ -55,7 +54,7 @@ public class HearingRecordingSegmentScenarios extends BaseTest {
             .statusCode(200)
             .body("folder-name", equalTo(FOLDER))
             .body("filenames", hasSize(1))
-            .body("filenames[0]", equalTo(FILE_NAME));
+            .body("filenames[0]", equalTo(fileName));
     }
 
     @Test
